requirements.txt
flask_restx
pillow
numpy
